refactor(TextLink): client/配下をフラット配置に整理する
client/components/UpwardLink/ という二重ネストを解消し、client/ 直下に
フラットに配置する。TextLink配下にはhookが存在せず、components/という
下位区分自体に意味がなかったため。UpwardLink専用に切られていた
stories/もコンポーネント直下のstories/へ統合する。

判断基準はCLAUDE.mdの「client 境界が必要なモジュールは client/ に
閉じ込める」節を参照。
