--- packages/smarthr-ui/src/components/TextLink/client/components/UpwardLink/index.ts
@@ -1,1 +1,0 @@
-export { UpwardLink } from './UpwardLink'
