+++ packages/smarthr-ui/src/components/TextLink/stories/TextLink.stories.tsx
@@ -3,7 +3,7 @@ import { action } from 'storybook/actions'
 import { FaArrowRightIcon, FaCircleQuestionIcon, OpenInNewTabIcon } from '../../Icon'
 import { Table, Th } from '../../Table'
 import { TextLink } from '../TextLink'
-import { UpwardLink } from '../client/components'
+import { UpwardLink } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
 import type { ComponentPropsWithoutRef } from 'react'
